requirement.txt
django
pillow
django-ckeditor
django-hitcount
django-import-export